Fix build warnings for MV-HEVC, Alpha & SCC and Code cleanup

+++ source/encoder/nal.cpp
@@ -114,7 +114,7 @@ void NALList::serialize(NalUnitType nalUnitType, const Bitstream& bs, int layerI
      * nuh_reserved_zero_6bits  6-bits
      * nuh_temporal_id_plus1    3-bits */
     out[bytes++] = (uint8_t)nalUnitType << 1;
-    out[bytes++] = (layerId << 3) | (temporalID);
+    out[bytes++] = (uint8_t)((layerId << 3) | (temporalID));
 
     /* 7.4.1 ...
      * Within the NAL unit, the following three-byte sequences shall not occur at
